LPS-71583 Separate Counters and Optimizers from blogs to its own module
