+++ adaptive-media-document-library-web/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Adaptive Media Document Library Web
+Bundle-SymbolicName: com.liferay.adaptive.media.document.library.web
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Adaptive Media
+-dsannotations-options: inherit
