+++ adaptive-media-document-library-web/build.gradle
@@ -2,10 +2,10 @@ sourceCompatibility = "1.8"
 targetCompatibility = "1.8"
 
 dependencies {
-	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
-	provided group: "com.liferay", name: "com.liferay.document.library.api", version: "3.2.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:adaptive-media:adaptive-media-api")
 	provided project(":apps:adaptive-media:adaptive-media-image-api")
+	provided project(":apps:adaptive-media:adaptive-media-web")
 }
