+++ adaptive-media-document-library-web/src/main/java/com/liferay/adaptive/media/document/library/web/internal/counter/DLAdaptiveMediaImageCounter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.document.library.internal.repository.counter;
+package com.liferay.adaptive.media.document.library.web.internal.counter;
 
 import com.liferay.adaptive.media.image.constants.AdaptiveMediaImageConstants;
 import com.liferay.adaptive.media.image.counter.AdaptiveMediaImageCounter;
